backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database and ORM
neo4j==5.14.1
chromadb==0.4.18
sqlalchemy==2.0.23
alembic==1.12.1

# AI and ML
google-generativeai==0.3.2
openai==1.3.7
langchain==0.0.350
langchain-google-genai==0.0.5
sentence-transformers==2.2.2
huggingface-hub==0.16.4

# Document processing
pymupdf==1.23.8
python-docx==1.1.0
pandas==2.1.3
openpyxl==3.1.2
python-pptx==0.6.23

# Text processing and NLP
spacy==3.7.2
nltk==3.8.1
textract==1.6.5

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0 